added configuration reading from file

+++ configuration/Configuration.hpp
@@ -9,7 +9,7 @@ namespace am {
 			size_t AffinityThreshold;
 			size_t MinPixelsForObject;
 			uint8_t PixelStep;
-			size_t CalculationTimeLimit;
+			double CalculationTimeLimit;
 			size_t IdleTimeout;
 		};
